🎨 style: orderDetailInfo 컴포넌트 생성 / order style 분리 / ui 부분 수정
