+++ src/styles/orderStyle.tsx
@@ -1,0 +1,128 @@
+import styled from "styled-components";
+
+export const OrderDetail = styled.main`
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+
+  h2 {
+    font-size: 22px;
+    font-weight: 600;
+    margin-bottom: 20px;
+  }
+
+  .progress_wrap {
+    gap: 82px;
+    margin: 20px;
+  }
+
+  .progress_and_text {
+    display: flex;
+    gap: 10px;
+    align-items: center;
+    flex-direction: column;
+    z-index: 10;
+  }
+
+  div {
+    display: flex;
+    text-align: center;
+  }
+
+  p {
+    font-weight: 600;
+    font-size: 15px;
+  }
+
+  .progress_wrap_article {
+    display: flex;
+    justify-content: center;
+    position: relative;
+  }
+`;
+
+export const OrderDetArticle = styled.article`
+  display: flex;
+  width: 100%;
+  padding: 10px;
+  align-items: center;
+  justify-content: space-around;
+  z-index: 10;
+  position: relative;
+
+  .img_and_info {
+    display: flex;
+    gap: 30px;
+  }
+
+  .product_info_wrap {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: start;
+  }
+
+  .product_title {
+    font-size: 20px;
+  }
+
+  .product_nickname {
+    font-size: 14px;
+  }
+
+  .btn_end {
+    padding: 10px 20px;
+    background-color: var(--color-main);
+    color: white;
+    font-weight: 500;
+  }
+`;
+
+export const ProductImg = styled.img`
+  width: 200px;
+  height: 200px;
+  border-radius: 15px;
+  object-fit: contain;
+`;
+
+export const ProgressBarWrap = styled.div`
+  display: flex;
+  gap: 30px;
+  margin: 20px auto;
+  top: 21px;
+  position: absolute;
+`;
+
+interface OrderProgressBarProps {
+  status?: string;
+}
+
+export const OrderProgressBar = styled.div<OrderProgressBarProps>`
+  margin-bottom: 30px;
+  width: 105px;
+  height: 5px;
+  background-color: #d8d7d7;
+
+  &:first-child {
+    background-color: ${(props) =>
+      props.status === "입금 대기" ? "#d8d7d7" : "#3EABFA"};
+  }
+
+  &:nth-child(2) {
+    background-color: ${(props) =>
+      props.status === "배송 시작" || props.status === "거래 종료"
+        ? "#3EABFA"
+        : "#d8d7d7"};
+  }
+
+  &:nth-child(3) {
+    background-color: ${(props) =>
+      props.status === "거래 종료" ? "#3EABFA" : "#d8d7d7"};
+  }
+`;
+
+// 주문 정보
+export const OrderInfo = styled.article`
+  display: flex;
+  flex-direction: column;
+`;
